todo_app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==4.2
django-extensions==3.2.3
djangorestframework==3.15.2
dnspython==2.7.0
eventlet==0.39.0
executing==2.2.0
greenlet==3.1.1
iniconfig==2.0.0
ipython==8.32.0
jedi==0.19.2
kombu==5.4.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
